requirements.txt
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7
pillow==10.0.0
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
streamlit==1.28.0
requests==2.31.0
plotly==5.17.0
pandas==2.1.1